+++ trixnity-crypto-core/src/webMain/kotlin/de/connect2x/trixnity/crypto/core/hmacSha256.kt
@@ -1,0 +1,39 @@
+package de.connect2x.trixnity.crypto.core
+
+import createHmac
+import io.ktor.util.*
+import js.array.jsArrayOf
+import js.buffer.toByteArray
+import js.json.rawJSON
+import js.objects.unsafeJso
+import js.typedarrays.Uint8Array
+import js.typedarrays.toByteArray
+import js.typedarrays.toInt8Array
+import js.typedarrays.toUint8Array
+import web.crypto.*
+import kotlin.js.toJsString
+
+actual suspend fun hmacSha256(key: ByteArray, data: ByteArray): ByteArray {
+    return if (PlatformUtils.IS_BROWSER) {
+        val crypto = crypto.subtle
+        val hmacKey = crypto.importKey(
+            format = KeyFormat.raw,
+            keyData = key.fastToUint8Array(),
+            algorithm = unsafeJso<HmacImportParams> {
+                name = "HMAC"
+                hash = "SHA-256".toJsString()
+            },
+            extractable = false,
+            keyUsages = jsArrayOf(KeyUsage.sign)
+        )
+        crypto.sign(
+            algorithm = "HMAC",
+            key = hmacKey,
+            data = data.fastToUint8Array()
+        ).toByteArray()
+    } else {
+        val hmac = createHmac(algorithm = "sha256", key = key.toUint8Array())
+        hmac.update(data.toUint8Array())
+        Uint8Array(hmac.digest()).toByteArray()
+    }
+}