+++ trixnity-libolm/src/jsMain/kotlin/de/connect2x/trixnity/libolm/utils.kt
@@ -1,7 +1,0 @@
-package de.connect2x.trixnity.libolm
-
-fun <T> rethrow(block: () -> T): T = try {
-    block()
-} catch (error: Throwable) {
-    throw OlmLibraryException(error.message?.substringAfter("OLM."), error)
-}